+++ tests/test_lab_dev/test_circuit_components_utils/test_b_to_ps.py
@@ -1,4 +1,4 @@
-# Copyright 2024 Xanadu Quantum Technologies Inc.
+# Copyright 2025 Xanadu Quantum Technologies Inc.
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,126 +12,48 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for BtoCH."""
+"""Tests for BtoPS."""
 
 # pylint: disable=fixme, missing-function-docstring, pointless-statement
 
 import numpy as np
 import pytest
 
-from mrmustard import math
-from mrmustard.lab_dev import DM, BtoCH, Identity, Ket
-from mrmustard.physics.bargmann_utils import wigner_to_bargmann_rho
-from mrmustard.physics.gaussian_integrals import complex_gaussian_integral_2
-from mrmustard.physics.triples import displacement_map_s_parametrized_Abc
+from mrmustard import math, settings
+from mrmustard.lab_dev import BtoPS, Dgate, Ket
+from mrmustard.physics.wigner import wigner_discretized
 
 
-class TestBtoCH:
+class TestBtoPS:
     r"""
-    Tests for the ``BtoCH`` class.
+    Tests for the ``BtoPS`` class.
     """
 
     modes = [(0,), (1, 2), (7, 9)]
-    s = [0, -1, 1]
+    s = [0, -0.9, 1]
 
     @pytest.mark.parametrize("modes,s", zip(modes, s))
     def test_init(self, modes, s):
-        dsmap = BtoCH(modes, s)
-        assert dsmap.name == "BtoCH"
-        assert dsmap.modes == modes
+        bw = BtoPS(modes, s)
+        assert bw.name == "BtoPS"
+        assert bw.modes == modes
+        assert bw.parameters.s.value == s
 
-    def test_adjoint(self):
-        b_to_ch = BtoCH(0, 0)
-        adjoint_BtoCH = b_to_ch.adjoint
+    @pytest.mark.parametrize("hbar", [1.0, 2.0, 3.0])
+    def test_application(self, hbar):
+        settings.HBAR = hbar
+        state = Ket.random((0,)) >> Dgate(0, x=2, y=0.1)
 
-        bras = b_to_ch.wires.bra.indices
-        kets = b_to_ch.wires.ket.indices
-        assert adjoint_BtoCH.ansatz == b_to_ch.ansatz.reorder(kets + bras).conj
-        assert adjoint_BtoCH.wires == b_to_ch.wires.adjoint
-        assert adjoint_BtoCH.parameters.s == b_to_ch.parameters.s
+        dm = math.sum(state.to_fock(100).dm().ansatz.array, axis=0)
+        vec = np.linspace(-5, 5, 100)
+        wigner, _, _ = wigner_discretized(dm, vec, vec)
 
-    def test_dual(self):
-        b_to_ch = BtoCH(0, 0)
-        dual_BtoCH = b_to_ch.dual
-
-        ok = b_to_ch.wires.ket.output.indices
-        ik = b_to_ch.wires.ket.input.indices
-        ib = b_to_ch.wires.bra.input.indices
-        ob = b_to_ch.wires.bra.output.indices
-        assert dual_BtoCH.ansatz == b_to_ch.ansatz.reorder(ib + ob + ik + ok).conj
-        assert dual_BtoCH.wires == b_to_ch.wires.dual
-        assert dual_BtoCH.parameters.s == b_to_ch.parameters.s
-
-    def test_inverse(self):
-        b_to_ch = BtoCH(0, 0)
-        inv_BtoCH = b_to_ch.inverse()
-        assert (b_to_ch >> inv_BtoCH).ansatz == (Identity(0) @ Identity(0).adjoint).ansatz
-
-    def test_representation(self):
-        ansatz = BtoCH(modes=0, s=0).ansatz
-        A_correct, b_correct, c_correct = displacement_map_s_parametrized_Abc(s=0, n_modes=1)
-        assert math.allclose(ansatz.A[0], A_correct)
-        assert math.allclose(ansatz.b[0], b_correct)
-        assert math.allclose(ansatz.c[0], c_correct)
-
-        ansatz2 = BtoCH(modes=(5, 10), s=1).ansatz
-        A_correct, b_correct, c_correct = displacement_map_s_parametrized_Abc(s=1, n_modes=2)
-        assert math.allclose(ansatz2.A[0], A_correct)
-        assert math.allclose(ansatz2.b[0], b_correct)
-        assert math.allclose(ansatz2.c[0], c_correct)
-
-    def testBtoCH_contraction_with_state(self):
-        # The init state cov and means comes from the random state 'state = Gaussian(1) >> Dgate([0.2], [0.3])'
-        state_cov = np.array([[0.32210229, -0.99732956], [-0.99732956, 6.1926484]])
-        state_means = np.array([0.4, 0.6])
-        A, b, c = wigner_to_bargmann_rho(state_cov, state_means)
-        state = DM.from_bargmann(modes=(0,), triple=(A, b, c))
-        state_bargmann_triple = state.bargmann_triple()
-
-        # get new triple by right shift
-        state_after = state >> BtoCH(modes=(0,), s=0)
-        A1, b1, c1 = state_after.bargmann_triple(batched=True)
-
-        # get new triple by contraction
-        Ds_bargmann_triple = displacement_map_s_parametrized_Abc(s=0, n_modes=1)
-        A2, b2, c2 = complex_gaussian_integral_2(
-            state_bargmann_triple, Ds_bargmann_triple, idx1=[0, 1], idx2=[1, 3]
+        Wigner = (state >> BtoPS([0], s=0)).ansatz
+        X, Y = np.meshgrid(vec / np.sqrt(2 * settings.HBAR), vec / np.sqrt(2 * settings.HBAR))
+        Z = np.array([X - 1j * Y, X + 1j * Y]).transpose((1, 2, 0))
+        assert math.allclose(
+            np.real(Wigner(Z)) / (2 * settings.HBAR),
+            np.real(wigner.T),
+            atol=1e-6,
         )
-
-        assert math.allclose(A1, A2)
-        assert math.allclose(b1, b2)
-        assert math.allclose(c1, c2)
-
-        # The init state cov and means comes from the random state 'state = Gaussian(2) >> Dgate([0.2], [0.3])'
-        state_cov = np.array(
-            [
-                [0.77969414, 0.10437996, 0.72706741, 0.29121535],
-                [0.10437996, 0.22846619, 0.1211067, 0.45983868],
-                [0.72706741, 0.1211067, 1.02215481, 0.16216756],
-                [0.29121535, 0.45983868, 0.16216756, 2.10006],
-            ]
-        )
-        state_means = np.array([0.28284271, 0.0, 0.42426407, 0.0])
-        A, b, c = wigner_to_bargmann_rho(state_cov, state_means)
-        state = DM.from_bargmann(modes=(0, 1), triple=(A, b, c))
-        state_bargmann_triple = state.bargmann_triple()
-
-        # get new triple by right shift
-        state_after = state >> BtoCH(modes=(0, 1), s=0)
-        A1, b1, c1 = state_after.bargmann_triple(batched=True)
-
-        # get new triple by contraction
-        Ds_bargmann_triple = displacement_map_s_parametrized_Abc(s=0, n_modes=2)
-        A2, b2, c2 = complex_gaussian_integral_2(
-            state_bargmann_triple,
-            Ds_bargmann_triple,
-            idx1=[0, 1, 2, 3],
-            idx2=[2, 3, 6, 7],
-        )
-
-        assert math.allclose(A1, A2)
-        assert math.allclose(b1, b2)
-        assert math.allclose(c1, c2)
-
-        psi = Ket.random([0])
-        assert math.allclose((psi >> BtoCH(0, 1)).ansatz([0, 0]), [1.0])
+        settings.HBAR = 1.0
